Interfaccia Heap

+++ pkg/heap/heap.go
@@ -5,7 +5,7 @@ type Heap interface {
 		Inserisce un elemento nello heap.
 		Precondizione: non deve esistere nodeID già all'interno dello heap
 	*/
-	Insert(nodeID int, distance int)
+	Insert(nodeID int, distance float64)
 	/*
 		Estrae il nodo con distanza minima.
 		Precondizione: heap non deve essere vuoto
@@ -15,7 +15,7 @@ type Heap interface {
 		Diminusce il valore nello heap di un nodo esistente
 		Precondizione: la nuova distanza deve essere minore e il nodo deve esistere nello heap
 	*/
-	DecreaseKey(nodeID int, newDistance int)
+	DecreaseKey(nodeID int, newDistance float64)
 	/*
 		Stampa la struttura dello heap
 	*/
